+++ source/Databricks/source/SqlStatementExecution/DatabricksSqlWarehouseQueryExecutor.cs
@@ -91,6 +91,52 @@ public class DatabricksSqlWarehouseQueryExecutor
         }
     }
 
+    /// <summary>
+    /// Asynchronously executes a parameterized SQL query on Databricks and streams the result back as a collection of strongly typed objects.
+    /// </summary>
+    /// <param name="statement">The SQL query to be executed, with collection of <see cref="QueryParameter"/> parameters.</param>
+    /// <typeparam name="T">Target type</typeparam>
+    /// <returns>An asynchronous enumerable of <typeparamref name="T"/> representing the result of the query.</returns>
+    /// <remarks>
+    /// This is an experimental feature and may be removed in a future version.
+    /// <br/><br/>
+    /// Requirements for <typeparamref name="T"/>:<br/>
+    /// - Must be a reference type<br/>
+    /// - Must have a public constructor with parameters matching the columns in the result set<br/>
+    /// - Must only have a single constructor<br/>
+    /// - Must be annotated with <see cref="ArrowFieldAttribute"/> to indicate the order of the constructor parameters
+    /// </remarks>
+    public virtual async IAsyncEnumerable<T> ExecuteStatementAsync<T>(DatabricksStatement statement)
+        where T : class
+    {
+        var strategy = new StronglyTypedApacheArrowFormat(_options);
+        var request = strategy.GetStatementRequest(statement);
+        var response = await request.WaitForSqlWarehouseResultAsync(_httpClient, StatementsEndpointPath);
+
+        if (_httpClient.BaseAddress == null) throw new InvalidOperationException();
+
+        if (response.manifest.total_row_count <= 0)
+        {
+            yield break;
+        }
+
+        foreach (var chunk in response.manifest.chunks)
+        {
+            var uri = StatementsEndpointPath +
+                      $"/{response.statement_id}/result/chunks/{chunk.chunk_index}?row_offset={chunk.row_offset}";
+            var chunkResponse = await _httpClient.GetFromJsonAsync<ManifestChunk>(uri);
+
+            if (chunkResponse?.external_links == null) continue;
+
+            await using var stream = await _externalHttpClient.GetStreamAsync(chunkResponse.external_links[0].external_link);
+
+            await foreach (var row in strategy.ExecuteAsync<T>(stream))
+            {
+                yield return row;
+            }
+        }
+    }
+
     private async IAsyncEnumerable<dynamic> DoExecuteStatementAsync(DatabricksStatement statement, Format format)
     {
         var strategy = format.GetStrategy(_options);